--- src/test/java/com/gradle/develocity/bamboo/UtilsTest.java
@@ -1,57 +1,0 @@
-package com.gradle.develocity.bamboo;
-
-import com.atlassian.bamboo.v2.build.BuildContext;
-import com.atlassian.bamboo.vcs.configuration.PlanRepositoryDefinition;
-import com.atlassian.bamboo.vcs.configuration.VcsLocationDefinition;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Sets;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-
-import java.util.Collections;
-import java.util.Set;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-class UtilsTest {
-
-    private final BuildContext buildContext = Mockito.mock(BuildContext.class);
-
-    @Test
-    void vcsRepoUrlsNoRepos() {
-        when(buildContext.getRelevantRepositoryIds())
-            .thenReturn(Collections.emptySet());
-
-        Set<String> repoUrls = Utils.vcsRepoUrls(buildContext);
-
-        MatcherAssert.assertThat(repoUrls, Matchers.emptyIterable());
-    }
-
-    @Test
-    void vcsRepoUrls() {
-        when(buildContext.getRelevantRepositoryIds())
-            .thenReturn(Sets.newHashSet(1L, 2L, 3L));
-        ImmutableMap<Long, PlanRepositoryDefinition> repoMap = ImmutableMap.of(
-            1L, mockPlanRepo("http://foo.git", "repositoryUrl"),
-            2L, mockPlanRepo("bar", "github.repository"),
-            3L, mockPlanRepo("http://random", "random"));
-        when(buildContext.getVcsRepositoryMap()).thenReturn(repoMap);
-
-        Set<String> repoUrls = Utils.vcsRepoUrls(buildContext);
-
-        MatcherAssert.assertThat(repoUrls, Matchers.equalTo(Sets.newHashSet("http://foo.git", "https://github.com/bar")));
-    }
-
-    private PlanRepositoryDefinition mockPlanRepo(String url, String key) {
-        PlanRepositoryDefinition mockRepo = mock(PlanRepositoryDefinition.class);
-        VcsLocationDefinition mockLoc = mock(VcsLocationDefinition.class);
-        when(mockLoc.getConfiguration()).thenReturn(ImmutableMap.of(key, url));
-        when(mockRepo.getVcsLocation()).thenReturn(mockLoc);
-        return mockRepo;
-    }
-
-
-}
